Let users know if no code owners are defined yet
As explained in [1] we have a special bootstrapping behaviour for
code-owners if
- the code owners functionality is enabled (and hence code owner
  approvals are required) and
- no code owners are defined yet (branch doesn't contain any OWNERS
  file yet)

In this case to submit a change the change owner either needs a code
owner override or an approval from a project owner. Currently users are
not aware of this, and the frontend should inform the users about this,
e.g. show this info as a message to the user when they try to get code
owners suggested. Since there are no code owners defined yet, we cannot
suggest any code owners and hence the suggestion is always empty in this
situation.

The Get Code Owner Branch Config REST endpoint was updated to include a
flag for this [2].

[1] https://gerrit-review.googlesource.com/c/plugins/code-owners/+/286639/2/resources/Documentation/user-guide.md
[2] https://gerrit-review.googlesource.com/c/plugins/code-owners/+/286473/2/resources/Documentation/rest-api.md    

Change-Id: Ia09cc5f43e1ed0684f4b443932356dd26107982d