AbstractRequiredApprovalConfigTest: Improve readability of 2 tests
1. testGetFromGlobalPluginConfig:
This method in the base class verified that "Code-Review+2" was
configured, but this configuration was done in the child classes via an
annotation. Due to this just looking at the
testGetFromGlobalPluginConfig() method it's unclear how the assertions
can work. To understand the assertions one must look at the calling
code. This is bad. Inline the method to improve the test readability,
even if it means that a small amount of code is being duplicated. After
doing this we can now use a more realistic label for the
getFromGlobalPluginConfig() test in OverrideApprovalConfigTest.

2. testCannotGetFromGlobalPluginConfigIfConfigIsInvalid:
Again the assertion only makes sense if you know the calling code.
Provide the invalid value as input, so that the assertion become more
readable.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I22a639cb878787fff8134ec4f3fcf397d21a1a20