AbstractRequiredApprovalConfig: Merge the get methods
Currently there is one get method to read the config from a plugin
config and one get method to read the config from the global config.
Merge boths methods into a single method that first reads the config
from the plugin config and then falls back to reading the config from
the global config if no value was found. This is more consistent with
the other methods for reading config values that all contain the
fallback logic.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I488f0eeafc193fd583b4c9dfca78403bfb7d9f22