Read all required approvals that are configured
We want to allow configuring multiple override approvals. As a first
step towards this we extend the config backend to read all configured
required approvals so that they get available for upper layers (namely
CodeOwnersPluginConfiguration).

With this change we still only support a single override approval. If
multiple override approvals are configured we use the last one. This is
the same behavior as before this change, since also JGit uses the last
configured value if a single config value is read, but multiple values
are configured.

Follow-up changes will adapt the upper layers to support multiple
override approvals.

Change-Id: I2a4387340e0336ea3ebe599b6e0fce913fa3689a
Signed-off-by: Edwin Kempin <ekempin@google.com>