Test behaviour with multiple configured required approvals
If in a JGit config file multiple values are configured for a key, but
only a single value is read (by Config#getString(String, String,
String)) the last value is returned by JGit. Cover this behaviour for
required approvals by tests so that we do not break this behavior when
we are going to support multiple values as override approval, but not as
required approval (to be implemented by follow-up changes).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I437144bdcf054aa145adfa963f77c9e094771a55