Support configuring multiple override approvals
With this change it is now possible to configure multiple approvals that
count as override for the code owners submit check. If multiple
approvals are configured, any of them is sufficient to override the code
owners submit check.

The configured override label is already returned to clients from the
Get Code Owner Branch Config REST endpoint and the Get Code Owner
Project Config REST endpoint, see the 'override_approval' field in the
respective response entity. Currently the 'override_approval' field
contains a RequiredApprovalInfo entity, but since we support multiple
override labels now we change this to a list of RequiredApprovalInfo
entities. This is an incompatible change that breaks the frontend.
Before submitting this change we must figure out how to adapt the
frontend so that it's not breaking.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I39e09c27ed4b52f60c460fe1ba1306e7073fe9fa