PathCodeOwners: Add an AutoValue class to represent the result
PathCodeOwners resolve a code owner config for a path. Add an AutoValue
class to represent the result of resolving the code owner config and
move the get() and ignoreParentCodeOwners() into it, as they operate on
the resolved code owner config. Since in the context of the new
AutoValue class the meaning of the get() method is not that clear,
rename it to getPathCodeOwners().

Having a dedicated result class will allow us to add further properties
to in in follow-up changes (e.g. a flag to tell whether there have been
unresolved imports).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ibd6e390b91ee15427c7da730bec1c5db9c209c35