PathCodeOwnersResult: Add flag that is being set if there are unresolved imports
Unresolved imports are silently ignored. However the fact that there are
unresolved imports is relevant for fallback code owners.

Fallback code owners should only apply if there are no code owners
defined for a path, but not if there are no code owners because an
import couldn't be resolved. In the latter case the intention was to
define code owners, and hence the fallback code owners should not apply.
Making use of the new flag in PathCodeOwnersResult for fallback code
owners will be implemented in a follow-up change.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Iad879cbddb07daba5c3741e3b380b3b7951cb711