Do not apply fallback code owners if there is a non-resolvable import
Fallback code owners should only apply if there are no code owners
defined for a path, but not if there are no code owners because an
import couldn't be resolved. In the latter case the intention was to
define code owners, and hence the fallback code owners should not apply.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I75ea58af0d28d86390ddbf663e3b32ea182920be