Add a documentation page with recommendations for the plugin configuration
The code-owners plugin comes with many configuration options and it's
hard for admins / project owners to understand which parameters they
should set for which purpose. We already have the setup guide that
documents the required configuration in detail, but beyond that admins /
project owners are required to dig through the descripions of the config
parameters themselves. To improve this a little, we add a new
documentation page with recommendations for the plugin configuration. At
this point it's still pretty empty, but follow-up changes will add more
contents.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I1cdfc0d804dfe5eb1a439fc592fa443434320675