CodeOwnerApprovalCheck: Allow to check for an account which paths it owns
When a code owner votes on the code owner label, we want to write a
change message that lists the files/paths that were approved by the user
by applying the vote. To write this message we need to know which of the
files/paths in the change are owned by the user. Instead of
reimplementing logic for this, we extend CodeOwnerApprovalCheck so that
it can compute the code owner statuses for all files/paths of a change
assuming that there is only an approval of a given user. Any file/path
that gets approved by this approval is owned by the user. Making use of
the new getFileStatusesForAccount method in CodeOwnerApprovalCheck to
write the change message on post review will be implemented by a
follow-up change.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I10551f22a36372beeea560fe357e86370fc9c756