On vote include approved paths into the change message
When a code owner votes on the code owner label, include into the change
message which files/paths:

* are approved now
* are no longer approved
* are still approved

This makes makes the effect on voting on the code owner label more
transparent to users.

In a follow-up change we may also extend the change message if a code
owner override is applied.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Idd11f2ea6021384a92f0a5588088e5cfda20a6c0