On vote include approved paths into the change message
When a code owner votes on the code owner label, include into the change
message which files/paths:

* are approved now
* are no longer approved
* are still approved

This makes makes the effect on voting on the code owner label more
transparent to users.

The paths in the change message are absolute file paths. They do not
have a leading '/' so that they match the paths in the file list on the
change screen which also do not have a leading '/'.

Example change message:

"John Doe is a code owner. By voting Code-Review+1 the following paths
are now approved by John Doe:
* foo/bar.baz
* foo/baz.bar"
* foo/bar/baz.foo"

Please note that the change message will contain old and new paths, e.g.
if a file is renamed, both the old and new path will be included
(assuming that they are both owned by the user).

In a follow-up change we may also extend the change message if a code
owner override is applied.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Idd11f2ea6021384a92f0a5588088e5cfda20a6c0