Limit the number of paths that are included in change messages
We must ensure that the change message, that is posted on review and in
which we include the approved files, doesn't grow too big for large
changes that touch many files. This is why we limit the number of paths
that are included in the change message. By default we include at most
100 paths, but we allow to configure this limit.

Example change message if the limit is exceeded:
"By voting Code-Review+1 the following files are now code-owner
approved by John Doe:
* bar/baz.foo
* bar/foo.baz
(2 more files)"

By setting the limit to 0, including affected paths into change messages
can also be disabled.

Change-Id: I5cbaec24ef54973f3dfea7e05091b80d5f9df369
Signed-off-by: Edwin Kempin <ekempin@google.com>