Allow to define and configure override label in the same commit
When a code-owners.config file is pushed we validate the override
configuration in it. This includes checking if the configured label
exists for the project. If the label didn't exist yet, but was
configured in the commit that is being pushed, we didn't take it into
account and wrongly rejected the commit, saying that the label doesn't
exist (because we read the labels from the head revision of
refs/meta/config and not from the commit that was being pushed). This is
fixed now, by reading the available labels from the commit that is being
pushed.

Change-Id: Ifecaca554605c123c62c347b54cb873f7cc7e09d
Signed-off-by: Edwin Kempin <ekempin@google.com>