Allow to ignore self approvals for required approval
If a label has ignoreSelfApproval set to true, self approvals from the
uploader are ignored. However so far this flag was ignored for code
owner approvals. This means the code owners check did pass if there was
an approval from the uploader, even if ignoreSelfApproval was set to
true. With this change now, any vote from the uploader is ignored for
the code owners check if the required label has ignoreSelfApproval set
to true. In this case the enableImplicitApprovals configuration is
ignored because there can never be an implicit approval of the uploader
if self approvals of the uploader are ignored.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I698a737bc2551e638c11b8cdd4ca3256fa3d45e0