Allow to ignore self approvals for overrides
If a label has ignoreSelfApproval set to true, self approvals from the
uploader are ignored. However so far this flag was ignored for code
owner overrides. This means the code owners check did pass if there was
an override from the uploader, even if ignoreSelfApproval was set to
true. With this change now, any override from the uploader is ignored
for the code owners check if the override label has ignoreSelfApproval
set to true.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I5c8486efc2f30152db83b56a04c7234e07ca68cf