Add test helper methods to set users as code owner
This removes duplicate code in tests and makes them more readable.

For now only the test of CodeOwnerApprovalCheck make use of the new
methods.

For some of the tests we now use a code owner config in the root
directory rather than a code owner config in the '/foo/' directory as
the call is a little simpler and for the tests it makes no difference.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I80b916f6af260f3ea3d1cb956054c81d661aab59