Validation: Do not reject imports that are added in the same commit
If a commit is validated that contains a code owner config with a new
import of a code owner config file that is added in the same commit, the
commit was wrongly rejected saying that the imported code owner config
was not resolveable. This was because we looked for the imported code
owner config in the HEAD revision of the branch. Instead we should look
it up from the commit from which also the importing code owner config
was loaded.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I32b0d8b9f3699e012e8c904e8602635e3a92a491