FindOwnersBackend: Fix incompatibility of handling globs
The find-owners backend promises to be compatible with the syntax of the
find-owners plugin, but currently the handling of '*' in globs differs.

The find-owners backend in the code-owners plugin currently handles '*'s
in globs like this:

* '*': matches any string that does not include slashes
* '**': matches any string, including slashes

This matches with how globs work in Java [1], however it is incompatible
with the find-owners plugin that handles '*'s in globs like this:

* '*'/'**': matches any string, including slashes

Change the find-owners backend in the code-owners plugin to be
compatible with the find-owners plugin. For this we simply replace any
single '*' character that appears in globs with '**' before we check if
the glob matches.

[1] https://mincong.io/2019/04/16/glob-expression-understanding/

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I1ddc8bf326bbcb8867ffd09eebc7a6739ad7d827