Support paging through code owners
Allow to skip code owners so that callers can page through the results.
Pagination requires a stable sort order, hence setting a seed to make
the sort order stable is required in this case.

TODO: sort order is not stable with same seed if limit is different,
using a different start changes the internal limit which makes the sort
order unstable, hence paging is not working (see test failures)

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I668bed6d7b54a49fd6a57e2401c0fac443673fb6