Count sticky approvals on old patch sets as code owner overrides
If an approval was applied on an old patch set which is still valid for
the current patch set because it is sticky we must consider it as code
owner override the same way as approvals that have been applied on the
current patch set.

The computation of sticky approvals is done by ApprovalInference which
we are now using by retrieving the approvals through ApprovalUtil.

The same issue was already fixed for code owner approvals by change
I2131162f5.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ifb7921336187c4b6d867ebe3397c9e1eed1d812d