PathCodeOwnersResult: Include details about unresolved imports
This will allow us to provide these details to host admins when they
need to investigate why certain code owner config files do not work as
expected. Requires further changes to expose this information to the
host admins.

The warnings about the unresolved imports do no longer need to be logged
right when they are detected, as the information about them is now
included into PathCodeOwnersResult which is fully logged, including the
information about the unresolved imports.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ie11efb52d2af41c66dabe25f67bfb9b9ff34479a