CodeOwnerResolver: Fix visibility checks if user is given
CodeOwnerResolver checks the visibility of the code owners, either for
the user provided by calling the forUser(IdentifiedUser) method or the
calling user.

So far if a user was provided by calling the forUser(IdentifiedUser)
method this user was only used for the account visibility check (see
canSee(AccountState) method) but not for the visibility checks of
secondary emails. The latter always used the current user. Fix this and
extend the tests to cover this case.

While we are here also update the javadocs that only spoke about
checking visibility for the calling user, while actually it could be a
different user, if one was provided by calling the
forUser(IdentifiedUser) method.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Icbb3627051cb546fc1823dfa85e9161d1f9f3e5e