CodeOwnerResolverResult: Include debug messages
Collect debug messages while resolving code owners so that we can make
them available to host admins if they need to investigate issues with
code owner config files that do not work as expected. Requires further
changes to expose this information to the host admins.

The debug messages are still written into the server logs on fine level
so that they are available in the logs when requests are traced.

Within CodeOwnerResolver the generation of debug messages is spread over
many methods. In order to be able to return them to the caller, we add
two new AutoValue classes (ResultWithMessages and
OptionalResultWithMessages) that wrap a result and also can contain
debug messages.

Adapt the CodeOwnerResolver tests to also verify the messages that
explain the result.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I12bb9d131815f461873e62d7a2b59b3e109b3055