Add REST endpoint to check if an email is a code owner of a path
This REST endpoint is intended to investigate code owner configurations
that do not work as intended. The response contains debug logs that may
point out issues with the code owner configuration. For example, with
this REST endpoint it is possible to find out why a certain email that
is listed as code owner in a code owner config file is ignored (e.g.
because it is ambiguous or because it belongs to an inactive account).

Since the returned debug logs may contain sensitive information, this
REST endpoint is only available to host admins that anyway have access
to this information.

The idea of adding this REST endpoint is to allow host admins to
investigate issues with code owner config files on their own, and hence
reduce support load for the Gerrit devs.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I477eda7b56f00a09894a5aa003cfd4eab4d9487b