Test toString methods of AutoValue classes
Strictly speaking the toString methods are not used, but having them is
useful for debugging.

The test add test coverage for the toString methods and verify that all
data is included in the string representation. E.g. such a test would
start failing if new data is added to the AutoValue class, but the
toString method was not adapted to include it.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6ef47f607b9dc4ef52570c2fc539f10bec3a10d2