CodeOwnersExceptionHook: Fix extracting message from exception cause
Instead of returning the message of the outer exception, return the
message of the cause. The cause should have a more detailed message,
e.g. if a code owner config file is invalid, it includes which line is
invalid.

For invalid code owner config files some information (project, branch)
was only available in the outer exception. Include this info into the
cause exception message too, so that we do not lose this info.

Change-Id: Ia26d880fd6574fe00cd76d3fcf5931b23f3e761b
Signed-off-by: Edwin Kempin <ekempin@google.com>