Use change number as seed for sorting when suggesting code owners for change
For suggesting code owners for paths in changes all requests should use
the same seed so that the sort order is stable across requests and
grouping of files with the same code owners works.

The initial idea was that the frontend generates a seed and then uses
the same seed for all requests that retrieve code owners for the change.
A disadvantage with this would be that the suggested code owners may
change whenever the change screen is reloaded (since reloading would
result in generating a new seed). To avoid this we will simply use the
change number as seed, if no seed was specified. This can be done on
server-side and there is also no need to adapt the frontend.

When listing code owners for a path in a branch we continue to use a
random seed if none was specified.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I1279e47666cca08c8d8ad616c4b9f073575c1abe