Add legacy REST endpoint for GetCodeOwnersForPathInChange REST endpoint
We need to change the response of the get code owners REST endpoints in
a backwards incompatible way. We can adapt the frontend so that it
understands both response formats however there are some callers of the
GetCodeOwnersForPathInChange REST endpoint that cannot be adapted in
this way. Hence for them the migration strategy is:

1. Add a GetCodeOwnersForPathInChangeLegacy REST endpoint that continues
   to support the old response (this change)
2. Switch the caller to using the GetCodeOwnersForPathInChangeLegacy
   REST endpoint the is added by this change
3. Change the GetCodeOwnersForPathInChange REST endpoint to the response
   (follow-up change)
4. Switch the caller to using the GetCodeOwnersForPathInChange REST
   endpoint with the new response
5. Drop the GetCodeOwnersForPathInChangeLegacy REST endpoint (second
   follow-up change)

The GetCodeOwnersForPathInChangeLegacy REST endpoind is just a wrapper
around the GetCodeOwnersForPathInChange REST endpoint that will preserve
the old response until all callers have been migrated.

The GetCodeOwnersForPathInBranch REST endpoint doesn't have callers that
need to be migrated, hence we don't need a legancy REST endpoint here.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ida668b052436038c83fc15762255c6247f788629