Make response of get code owners REST endpoints extensible
At the moment the get code owners REST endpoints return a list of
CodeOwnerInfo's. This doesn't allow us to add further fields to the
response in the future (e.g. we consider adding a
"is_owned_by_all_users" boolean field). To make the response extensible
we change it from returning List<CodeOwnerInfo> to returning
CodeOwnersInfo that has a "code_owners" field that contains the code
owners as List<CodeOwnerInfo>.

This is a backwards incompatible change and breaks existing callers of
this API. To avoid this the frontend was adapted to be able to
understand both response formats and other callers have been migrated to
use the code_owners.legacy REST endpoint instead which still returns the
old response format.

Change-Id: Ib0c4c8edf3952b48bccdc99baf77380a3b3f0c49
Signed-off-by: Edwin Kempin <ekempin@google.com>