Get code owners REST endpoints: Add option whether to resolve all users
If code ownership is assigned to all users, we currently fill up the
response with random users until the requested limit is reached. Add a
new request option that allows callers to control whether all users
should be resolved to random users.

Since the response now contains an explicit flag telling callers if a
path is owned by all users, resolving all users to random users may not
be wanted anymore.

By default all users are resolved to random users so that the current
behaviour is kept.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I8d05e936a32d2ad173dc8a868d352571cecdde9f