Allow to set a file extension for code owner config files via REST
So far setting a file extension for code owner config files required
fetching refs/meta/config, editing the code-owners.config file and
pushing back the updates to refs/meta/config. Using a REST endpoint for
this is more comfortable since it is a single command.

We want to make setting a file extension easier since this needs to be
done whenever an upstream project is imported that already contains code
owner config files.

Change-Id: Iaf7cfef8a64a7ce24ac667c2635527b9ed7f747e
Signed-off-by: Edwin Kempin <ekempin@google.com>