Extend change message when code owner override is applied
When a code owner override is applied, extend the change message that is
posted for the vote and say that this vote counts as code owner
override. This makes the effect of voting on the code owner override
label transparent to users. This is important since the code owner
override label is configurable and can have any name and users may not
be aware that a vote on this label counts as code owner override.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I6d073eb97c324080bf3f44629a5d92020b6fbcd5