FindOwnersBackend: Fix incompatibility of handling globs
The find-owners backend promises to be compatible with the syntax of the
find-owners plugin, but currently the handling of globs differs.

The find-owners plugin ignores subdirectories when checking if a glob
matches a file.

Examples:
1. the glob "*.md" in "/" matches "foo.md", "bar/foo.md" and
   "baz/bar/foo.md"
2. the glob "BUILD" in "/" matches "BUILD", "bar/BUILD" and
   "baz/bar/BUILD"
3. the glob "foo/**" in "/" matches "foo/README", "bar/foo/README" and
   "baz/bar/foo/README"
4. the glob "foo-[1-4].txt" in "/" matches "foo-1.txt", "bar/foo-1.txt"
   and "baz/bar/foo-1.txt"

This is because the find-owners plugin takes the glob and the filename
and then iterates over all parent directories [1,2] to see if the glob
matches the filename [3]. E.g. if the glob is "*.md" and the file path
is "baz/bar/foo.md" then it does the following checks:
a. Does the glob "*.md" match the file "./baz/bar/foo.md" (no)
b. Does the glob "*.md" match the file "./baz/foo.md" (no)
c. Does the glob "*.md" match the file "./foo.md" (yes)

To be compatible with this behaviour the code-owners plugin needs to
ignore subdirectories too when checking if a file matches a glob.
Instead of iterating over the parent directories for each file as it's
done in the find-owners plugin, we always do 2 glob matches, one for the
actual glob and one for "**/" + glob which matches all subdirectories,
e.g.:
I. Does the glob "*.md" match the file "baz/bar/foo.md" (no)
II. Does the glob "**/*.md" match the file "baz/bar/foo.md" (yes)

Doing II. only is not sufficient since "foo.md" would not be matched by
it.

Change I1ddc8bf32 was a first attempt to fix this by replacing "*" with
"**" in globs, but this only fixed case 1 (and not the cases 2-4).

[1] https://gerrit.googlesource.com/plugins/find-owners/+/refs/heads/master/src/main/java/com/googlesource/gerrit/plugins/findowners/OwnersDb.java#367
[2] https://gerrit.googlesource.com/plugins/find-owners/+/refs/heads/master/src/main/java/com/googlesource/gerrit/plugins/findowners/OwnersDb.java#396
[3] https://gerrit.googlesource.com/plugins/find-owners/+/refs/heads/master/src/main/java/com/googlesource/gerrit/plugins/findowners/OwnersDb.java#375

Bug: Issue 14008
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I04931606858511fb1c36c449a4ce3f2300eabecf