Remove equals/hashCode/toString methods from info classes
Mostly these were unused. Only one test assertion in
CodeOwnerConfigInfoSubject required some of the equals methods, but we
can rewrite this so that it no longer relies on the equals methods.

Updating equals/hashCode/toString can easily be forgotten if there are
no dedicated tests for them (e.g. in CodeOwnerConfigInfo they did not
consider the ignoreParentCodeOwners field which was not noticed). Since
they are not really needed and writing tests for them is a hassle (there
are utilities for this, but they are currently not available in upstream
Gerrit) we just drop them.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I74048bb0d1848ebb25b8c7fb93b76114d0bd7b45